@eltoder 2016-10-23
I've also have a change to do the same for arithmetic methods. I'll send over once this is merged.